Update README.md
Hi there! My name is Steven, and this is my portfolio containing all the work I have done as an analytics specialist and researcher based in Toronto. All the code that you see on this page are written from the ground up by myself, which includes the preparation, planning, brainstorming, and reporting of all the data related work I have done.